Provide improved mechanism to add GATK jars into bcbio-nextgen. Handles updating version information in manifest and supports academic and commerical downloads. Fixes #332

+++ bcbio/install.py
@@ -3,8 +3,10 @@
 Enables automated installation tool and in-place updates to install additional
 data and software.
 """
+import argparse
 import collections
 import contextlib
+import datetime
 from distutils.version import LooseVersion
 import os
 import shutil
@@ -15,7 +17,7 @@ import sys
 import requests
 import yaml
 
-from bcbio import utils
+from bcbio import broad, utils
 from bcbio.pipeline import genome
 from bcbio.variation import effects
 from bcbio.provenance import programs
@@ -27,6 +29,8 @@ REMOTES = {
     "genome_resources": "https://raw.github.com/chapmanb/bcbio-nextgen/master/config/genomes/%s-resources.yaml",
     "snpeff_dl_url": ("http://downloads.sourceforge.net/project/snpeff/databases/v{snpeff_ver}/"
                       "snpEff_v{snpeff_ver}_{genome}.zip")}
+
+Tool = collections.namedtuple("Tool", ["name", "fname"])
 
 def upgrade_bcbio(args):
     """Perform upgrade of bcbio to latest release, or from GitHub development version.
@@ -91,19 +95,17 @@ def _install_container_bcbio_system(datadir):
         if len(expose_vals) > 0 and pname not in expose_config["resources"]:
             expose_config["resources"][pname] = expose_vals
     with open(expose_file, "w") as out_handle:
-        yaml.dump(expose_config, out_handle, default_flow_style=False, allow_unicode=False)
+        yaml.safe_dump(expose_config, out_handle, default_flow_style=False, allow_unicode=False)
     return expose_file
 
 def _default_deploy_args(args):
-    flavors = {"minimal": "ngs_pipeline_minimal",
-               "full": "ngs_pipeline"}
-    toolplus = {"protected": {"bio_nextgen": ["gatk-protected"]},
-                "data": {"bio_nextgen": []}}
+    toolplus = {"data": {"bio_nextgen": []}}
     custom_add = collections.defaultdict(list)
     for x in args.toolplus:
-        for k, vs in toolplus[x].iteritems():
-            custom_add[k].extend(vs)
-    return {"flavor": flavors[args.tooldist],
+        if not x.fname:
+            for k, vs in toolplus.get(x.name, {}).iteritems():
+                custom_add[k].extend(vs)
+    return {"flavor": "ngs_pipeline_minimal",
             "custom_add": dict(custom_add),
             "vm_provider": "novm",
             "hostname": "localhost",
@@ -150,7 +152,7 @@ def upgrade_bcbio_data(args, remotes):
     _upgrade_genome_resources(s["fabricrc_overrides"]["galaxy_home"],
                               remotes["genome_resources"])
     _upgrade_snpeff_data(s["fabricrc_overrides"]["galaxy_home"], args, remotes)
-    if 'data' in args.toolplus:
+    if 'data' in set([x.name for x in args.toolplus]):
         gemini = os.path.join(os.path.dirname(sys.executable), "gemini")
         subprocess.check_call([gemini, "update", "--dataonly"])
 
@@ -229,19 +231,80 @@ def upgrade_thirdparty_tools(args, remotes):
     s["actions"] = ["install_biolinux"]
     s["fabricrc_overrides"]["system_install"] = args.tooldir
     s["fabricrc_overrides"]["local_install"] = os.path.join(args.tooldir, "local_install")
-    if "data" in args.toolplus:
-        _install_gemini(args.tooldir, _get_data_dir(), args)
+    manifest_dir = os.path.join(_get_data_dir(), "manifest")
+    _install_toolplus(args, manifest_dir)
     cbl = get_cloudbiolinux(remotes)
     sys.path.insert(0, cbl["dir"])
     cbl_deploy = __import__("cloudbio.deploy", fromlist=["deploy"])
     cbl_deploy.deploy(s)
     cbl_manifest = __import__("cloudbio.manifest", fromlist=["manifest"])
-    manifest_dir = os.path.join(_get_data_dir(), "manifest")
     print("Creating manifest of installed packages in %s" % manifest_dir)
     if os.path.exists(manifest_dir):
         for fname in os.listdir(manifest_dir):
-            os.remove(os.path.join(manifest_dir, fname))
+            if not fname.startswith("toolplus"):
+                os.remove(os.path.join(manifest_dir, fname))
     cbl_manifest.create(manifest_dir, args.tooldir)
+
+def _install_toolplus(args, manifest_dir):
+    """Install additional tools we cannot distribute, updating local manifest.
+    """
+    toolplus_manifest = os.path.join(manifest_dir, "toolplus-packages.yaml")
+    system_config = os.path.join(_get_data_dir(), "galaxy", "bcbio_system.yaml")
+    toolplus_dir = os.path.join(_get_data_dir(), "toolplus")
+    for tool in args.toolplus:
+        if tool.name == "data":
+            _install_gemini(args.tooldir, _get_data_dir(), args)
+        elif tool.name in set(["gatk", "mutect"]):
+            _install_gatk_jar(tool.name, tool.fname, toolplus_manifest, system_config, toolplus_dir)
+        elif tool.name in set(["protected"]):  # back compatibility
+            pass
+        else:
+            raise ValueError("Unexpected toolplus argument: %s %s" (tool.name, tool.fname))
+
+def _install_gatk_jar(name, fname, manifest, system_config, toolplus_dir):
+    """Install a jar for GATK or associated tools like MuTect.
+    """
+    if not fname.endswith(".jar"):
+        raise ValueError("--toolplus argument for %s expects a jar file: %s" % (name, fname))
+    if name == "gatk":
+        version = broad.get_gatk_version(fname)
+    elif name == "mutect":
+        version = broad.get_mutect_version(fname)
+    else:
+        raise ValueError("Unexpected GATK input: %s" % name)
+    store_dir = utils.safe_makedir(os.path.join(toolplus_dir, name, version))
+    shutil.copyfile(fname, os.path.join(store_dir, os.path.basename(fname)))
+    _update_system_file(system_config, name, {"dir": store_dir})
+    _update_manifest(manifest, name, version)
+
+def _update_manifest(manifest_file, name, version):
+    """Update the toolplus manifest file with updated name and version
+    """
+    if os.path.exists(manifest_file):
+        with open(manifest_file) as in_handle:
+            manifest = yaml.load(in_handle)
+    else:
+        manifest = {}
+    manifest[name] = {"name": name, "version": version}
+    with open(manifest_file, "w") as out_handle:
+        yaml.safe_dump(manifest, out_handle, default_flow_style=False, allow_unicode=False)
+
+def _update_system_file(system_file, name, new_kvs):
+    """Update the bcbio_system.yaml file with new resource information.
+    """
+    bak_file = system_file + ".bak%s" % datetime.datetime.now().strftime("%Y-%m-%d-%H-%M-%S")
+    shutil.copyfile(system_file, bak_file)
+    with open(system_file) as in_handle:
+        config = yaml.load(in_handle)
+    new_rs = {}
+    for rname, r_kvs in config.get("resources", {}).iteritems():
+        if rname == name:
+            for k, v in new_kvs.iteritems():
+                r_kvs[k] = v
+        new_rs[rname] = r_kvs
+    config["resources"] = new_rs
+    with open(system_file, "w") as out_handle:
+        yaml.safe_dump(config, out_handle, default_flow_style=False, allow_unicode=False)
 
 def _install_gemini(tooldir, datadir, args):
     """Install gemini layered on top of bcbio-nextgen, sharing anaconda framework.
@@ -291,20 +354,26 @@ def save_install_defaults(args):
             cur_config = yaml.load(in_handle)
     else:
         cur_config = {}
-    if args.tooldist not in "minimal":
-        cur_config["tooldist"] = args.tooldist
     if args.tooldir:
         cur_config["tooldir"] = args.tooldir
     cur_config["sudo"] = args.sudo
     cur_config["isolate"] = args.isolate
-    for attr in ["genomes", "aligners", "toolplus"]:
+    for attr in ["genomes", "aligners"]:
         if not cur_config.get(attr):
             cur_config[attr] = []
         for x in getattr(args, attr):
             if x not in cur_config[attr]:
                 cur_config[attr].append(x)
+    # toolplus -- save non-filename inputs
+    attr = "toolplus"
+    if not cur_config.get(attr):
+        cur_config[attr] = []
+    for x in getattr(args, attr):
+        if not x.fname:
+            if x.name not in cur_config[attr]:
+                cur_config[attr].append(x.name)
     with open(install_config, "w") as out_handle:
-        yaml.dump(cur_config, out_handle, default_flow_style=False, allow_unicode=False)
+        yaml.safe_dump(cur_config, out_handle, default_flow_style=False, allow_unicode=False)
 
 def add_install_defaults(args):
     """Add any saved installation defaults to the upgrade.
@@ -314,8 +383,6 @@ def add_install_defaults(args):
         return args
     with open(install_config) as in_handle:
         default_args = yaml.load(in_handle)
-    if default_args.get("tooldist") and args.tooldist == "minimal":
-        args.tooldist = str(default_args["tooldist"])
     if args.tools and args.tooldir is None:
         if "tooldir" in default_args:
             args.tooldir = str(default_args["tooldir"])
@@ -326,9 +393,10 @@ def add_install_defaults(args):
                              "work for future upgrades.")
     for attr in ["genomes", "aligners", "toolplus"]:
         for x in default_args.get(attr, []):
+            x = Tool(x, None) if attr == "toolplus" else str(x)
             new_val = getattr(args, attr)
             if x not in getattr(args, attr):
-                new_val.append(str(x))
+                new_val.append(x)
             setattr(args, attr, new_val)
     if "sudo" in default_args and not args.sudo is False:
         args.sudo = default_args["sudo"]
@@ -343,6 +411,22 @@ def get_defaults():
     with open(install_config) as in_handle:
         return yaml.load(in_handle)
 
+def _check_toolplus(x):
+    """Parse options for adding non-standard/commercial tools like GATK and MuTecT.
+    """
+    std_choices = set(["data"])
+    if x in std_choices:
+        return Tool(x, None)
+    elif "=" in x and len(x.split("=")) == 2:
+        name, fname = x.split("=")
+        fname = os.path.normpath(os.path.realpath(fname))
+        if not os.path.exists(fname):
+            raise argparse.ArgumentTypeError("Unexpected --toolplus argument for %s. File does not exist: %s"
+                                             % (name, fname))
+        return Tool(name, fname)
+    else:
+        raise argparse.ArgumentTypeError("Unexpected --toolplus argument. Expect toolname=filename.")
+
 def add_subparser(subparsers):
     parser = subparsers.add_parser("upgrade", help="Install or upgrade bcbio-nextgen")
     parser.add_argument("--tooldir",
@@ -354,7 +438,7 @@ def add_subparser(subparsers):
     parser.add_argument("-u", "--upgrade", help="Code version to upgrade",
                         choices=["stable", "development", "system", "skip"], default="skip")
     parser.add_argument("--toolplus", help="Specify additional tool categories to install",
-                        action="append", default=[], choices=["protected", "data"])
+                        action="append", default=[], type=_check_toolplus)
     parser.add_argument("--genomes", help="Genomes to download",
                         action="append", default=["GRCh37"],
                         choices=["GRCh37", "hg19", "mm10", "mm9", "rn5", "canFam3"])
@@ -367,10 +451,6 @@ def add_subparser(subparsers):
                         dest="sudo", action="store_false", default=True)
     parser.add_argument("--isolate", help="Created an isolated installation without PATH updates",
                         dest="isolate", action="store_true", default=False)
-    parser.add_argument("--tooldist",
-                        help="Type of tool distribution to install. Defaults to a minimum install.",
-                        default="minimal",
-                        choices=["minimal", "full"])
     parser.add_argument("--distribution", help="Operating system distribution",
                         default="",
                         choices=["ubuntu", "debian", "centos", "scientificlinux", "macosx"])
